+++ src/main/java/dta/sfmflow/datagen/ModBlockStateProvider.java
@@ -124,9 +124,9 @@ public class ModBlockStateProvider extends BlockStateProvider {
 		simpleBlockItem(ModBlocks.ENERGY_RELAY_BLOCK.get(), models().cubeAll("energy_relay_block", modLoc("block/energy_relay_block")));
 		
 		BlockModelBuilder signUpdaterModel = models().orientable("sign_updater_cable_block",
-				modLoc("block/sign_cable_side"), // Default side re-use
-				modLoc("block/sign_cable_front"), // Default front re-use
-				modLoc("block/sign_cable_side")); // Default top/back re-use
+				modLoc("block/sign_cable_side"), 
+				modLoc("block/sign_cable_front"), 
+				modLoc("block/sign_cable_side")); 
 		directionalBlockFacingNorth(ModBlocks.SIGN_UPDATER_CABLE_BLOCK.get(), signUpdaterModel);
 		simpleBlockItem(ModBlocks.SIGN_UPDATER_CABLE_BLOCK.get(), signUpdaterModel);